Add issubset and issuperset method to FeatureIDs object.
Add issubset and issuperset method to FeatureIDs object so it will be
easy to check if all feature IDs of a FeatureIDs object are at least all found
in another FeatureIDs object. This will be useful to check that after
lifting over regions/genes to another species, that the resulting
feature IDs are at least a subset of the original regions.