Add option to make deterministic rankings databases.
Add option to make deterministic rankings databases.
When deterministic_per_motif_or_track is set to true, the motif
or track name is used to initialize the random seed for ranking
scores for a certain motif or track. So in case you scored more
or less motifs or tracks, the ranking for a certain motif or
track will be the same between databases.