Make rankings per motif/track deterministic based on the current motif/track name when seed is set.
Make rankings per motif/track deterministic for tied scores based
on the current motif/track name when a seed is set.

This will result in the same ranking for a certain motif/track
between different cisTarget databases (with a different number
of motifs/tracks or differnt motifs/tracks, which would change
the position of the motif/track in the database).

The old behaviour would result in the same rankings cisTarget
database if a seed was set and the same motifs/tracks were used
between different runs.